+++ uk.ac.gda.devices.bssc.ui/src/uk/ac/gda/devices/bssc/ui/wizards/BSSCImportWizardPage.java
@@ -152,7 +152,7 @@ public class BSSCImportWizardPage extends WizardNewFileCreationPage {
 					tibi.setMolecularWeight(row.getCell(MOLECULAR_WEIGHT_COL_NO).getNumericCellValue());
 					tibi.setTimePerFrame(row.getCell(TIME_PER_FRAME_COL_NO).getNumericCellValue());
 					tibi.setFrames((int) row.getCell(FRAMES_COL_NO).getNumericCellValue()); 
-					tibi.setExposureTemperature((float) row.getCell(EXPOSURE_TEMP_COL_NO).getNumericCellValue()); 
+					tibi.setExposureTemperature(row.getCell(EXPOSURE_TEMP_COL_NO).getNumericCellValue()); 
 	
 					measurements.add(tibi);
 				} catch (Exception e) {
